string

+++ _print_number2.c
@@ -71,7 +71,7 @@ void print_array(int *array, int o)
  * @c: long int
  * Return: int
  */
-int count_size(long int c)
+int count_size(unsigned int c)
 {
 	int t;
